requirements.txt
pygame==2.5.2
pillow==10.1.0
flask==3.0.0 